Easy Python 1-File Discord Bot with basic commands
An easy to setup discord bot with basic commands that can be changed or extended everything run in the main.py file and even the Token can be if wanted put into that file, only requirements are Python installed and a already setup discord application and bot, all the coding stuff is in here. A perfect quick starter package for who want to make their first discord bot or those who make many of them and just don't wanna set everything up again everytime. Just put the exe file in a folder, run it (it's no virus it's just one of those self-extracting zips, I literally made it with Win-RAR) it will extract itself, then look for further instructions in the README.txt file! And Have fun!
(I think nobody will trust me when I say that it isn't a virus so I also put it the files in the Bot.zip too)